Improvements for function inject().
inject() is becoming essential to the pkg. Now, if the very first element passed to ... is either a list or a vector of length greater than 1, it is assumed that the child elements of ..1 should be used instead of ... Therefore, you can either pass args to inject() by parts (...) or as a whole (..1).