Collective changes to unit testing framework.
These changes were introduced at the same time the tests for class Atomic were written. This is because the framework needs to be as consistent as possible and thus, a lot of small changes here and there were introduced.

It would not be useful to list them all and justify them. They are simply maturing with the package.